// src/types/enums.ts
export enum TabEnum {
  Libraries = 'libraries',
  MyImages = 'myImages',
  Performance = 'performance',
}

export enum TypeEnum {
  SALE = 'SALE',
  RECRUIT = 'RECRUIT',
}

export enum LabelEnum {
  NEW = 'NEW',
  MARQUEE = 'MARQUEE',
  FEATURED = 'FEATURED',
  HOT = 'HOT',
  TOPUSED = 'TOPUSED',
  FAVORITE = 'FAVORITE',
}

export enum StatusEnum {
  ALL = 'ALL',
  SALE = 'SALE',
  RECRUIT = 'RECRUIT',
}

export enum SortEnum {
  Default = 'default',
  Newest = 'sortNewest',
  Oldest = 'sortOldest',
  MostInteractions = 'sortMostInteractions',
  MostFee = 'sortMostFee',
  MostESign = 'sortMostESign',
  FromAToZ = 'sortFromAToZ',
  FromZToA = 'sortFromZToA',
}

export enum TimeLineEnum {
  ONE_MONTH = 'oneMonth',
  ONE_QUARTER = 'oneQuarter',
  ONE_YEAR = 'oneYear',
}

export enum AvatarStatusOption {
  Waiting = 'waiting',
  Reviewed = 'reviewed',
  Rejected = 'rejected',
}

export enum AvatarActionOption {
  Display = 'display',
  Edit = 'edit',
  Delete = 'delete',
}